chore(file-upload): moved primitives to separate dirs
