+++ packages/components/file-upload/primitives/file-drop.spec.ts
@@ -2,8 +2,8 @@ import { Component } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { dispatchEvent, dispatchFakeEvent } from '@koobiq/cdk/testing';
+import { KbqFile } from '../file-upload';
 import { KbqFileDropDirective } from './file-drop';
-import { KbqFile } from './file-upload';
 
 export const createFile = (
     name: string,
