+++ packages/components/file-upload/primitives/index.ts
@@ -1,0 +1,2 @@
+export * from './file-drop';
+export * from './file-picker';
