+++ packages/components/file-upload/public-api.ts
@@ -1,6 +1,6 @@
-export * from './file-drop';
-export * from './file-picker';
 export * from './file-upload';
 export * from './file-upload.module';
 export * from './multiple-file-upload.component';
+export * from './primitives/file-drop';
+export * from './primitives/file-picker';
 export * from './single-file-upload.component';
